# src/swagger.yaml
openapi: 3.0.3
info:
  title: Scraawl API
  description: >- 
  version: 1.0.0
  contact: {}
servers:
  - url: 'http://localhost:8080/scraawl/api'
paths:
...
  /users/login:
    post:
      tags:
        
        - 🔐 Authentication
      summary: Login user
      description: >-
        The API endpoint allows users to log into the application securely and
        also returns an access token.


        When accessing this endpoint with valid login credentials, users will
        receive an access token in the response.


        Additionally, the API endpoint securely sets the access token within the
        browser/client httpOnly cookies for future authentication and
        authorization purposes.


        This functionality ensures a secure and efficient login process for
        users, providing them with an access token that can be used to
        authenticate subsequent API requests.


        By securely storing the access token in browser cookies, the endpoint
        enables automatic inclusion of the access token in future API requests,
        eliminating the need for users to manually manage and provide the token
        with each request.
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: test@123
                username:
                  type: string
                  example: doejohn
            example:
              password: test@123
              username: doejohn
      responses:
        '200':
          description: Login user
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '926'
            Date:
              schema:
                type: string
                example: Fri, 30 Jun 2023 06:44:29 GMT
            ETag:
              schema:
                type: string
                example: W/"39e-a7Fdn0Th2/i/29/rw1he3r/pebo"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            RateLimit-Limit:
              schema:
                type: string
                example: '500'
            RateLimit-Remaining:
              schema:
                type: string
                example: '494'
            RateLimit-Reset:
              schema:
                type: string
                example: '648'
            Set-Cookie:
              schema:
                type: string
                example: >-
                  refreshToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDllNzk1MjY0ZTdkYmEyOWI3MjY5MzYiLCJpYXQiOjE2ODgxMDc0NjksImV4cCI6MTY4ODk3MTQ2OX0.7or55Coj0K_UY4YpzYzkNR9hcOIhusxk9FRz7eMV8Ho;
                  Path=/; HttpOnly
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                        example: >-
                          eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDllNzk1MjY0ZTdkYmEyOWI3MjY5MzYiLCJlbWFpbCI6IndhamUuc2h1YmhhbTExMUBnbWFpbC5jb20iLCJ1c2VybmFtZSI6IndhamVzaHViaGFtIiwicm9sZSI6IkFETUlOIiwiaWF0IjoxNjg4MTA3NDY5LCJleHAiOjE2ODgxOTM4Njl9.40qfVrWi8x8fGBRB9JBVCVsdMYvlwzjIRpffz5z7DK4
                      refreshToken:
                        type: string
                        example: >-
                          eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDllNzk1MjY0ZTdkYmEyOWI3MjY5MzYiLCJpYXQiOjE2ODgxMDc0NjksImV4cCI6MTY4ODk3MTQ2OX0.7or55Coj0K_UY4YpzYzkNR9hcOIhusxk9FRz7eMV8Ho
                      user:
                        type: object
                        properties:
                          __v:
                            type: number
                            example: 0
                          _id:
                            type: string
                            example: 649e795264e7dba29b726936
                          avatar:
                            type: object
                            properties:
                              _id:
                                type: string
                                example: 649e795264e7dba29b726935
                              localPath:
                                type: string
                                example: ''
                              url:
                                type: string
                                example: https://via.placeholder.com/200x200.png
                          createdAt:
                            type: string
                            example: '2023-06-30T06:42:26.163Z'
                          email:
                            type: string
                            example: user.email@domain.com
                          isEmailVerified:
                            type: boolean
                            example: false
                          loginType:
                            type: string
                            example: EMAIL_PASSWORD
                          role:
                            type: string
                            example: ADMIN
                          updatedAt:
                            type: string
                            example: '2023-06-30T06:44:29.831Z'
                          username:
                            type: string
                            example: doejohn
                  message:
                    type: string
                    example: User logged in successfully
                  statusCode:
                    type: number
                    example: 200
                  success:
                    type: boolean
                    example: true
              examples:
                Login user:
                  value:
                    data:
                      accessToken: >-
                        eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDllNzk1MjY0ZTdkYmEyOWI3MjY5MzYiLCJlbWFpbCI6IndhamUuc2h1YmhhbTExMUBnbWFpbC5jb20iLCJ1c2VybmFtZSI6IndhamVzaHViaGFtIiwicm9sZSI6IkFETUlOIiwiaWF0IjoxNjg4MTA3NDY5LCJleHAiOjE2ODgxOTM4Njl9.40qfVrWi8x8fGBRB9JBVCVsdMYvlwzjIRpffz5z7DK4
                      refreshToken: >-
                        eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDllNzk1MjY0ZTdkYmEyOWI3MjY5MzYiLCJpYXQiOjE2ODgxMDc0NjksImV4cCI6MTY4ODk3MTQ2OX0.7or55Coj0K_UY4YpzYzkNR9hcOIhusxk9FRz7eMV8Ho
                      user:
                        __v: 0
                        _id: 649e795264e7dba29b726936
                        avatar:
                          _id: 649e795264e7dba29b726935
                          localPath: ''
                          url: https://via.placeholder.com/200x200.png
                        createdAt: '2023-06-30T06:42:26.163Z'
                        email: user.email@domain.com
                        isEmailVerified: false
                        loginType: EMAIL_PASSWORD
                        role: ADMIN
                        updatedAt: '2023-06-30T06:44:29.831Z'
                        username: doejohn
                    message: User logged in successfully
                    statusCode: 200
                    success: true
...